Extracted common types closes https://github.com/SoftUni-Internal/exam-systems-issues/issues/134
